src/main/java/com/sun/app/trades/TradeDAO.java
package com.sun.app.trades;

import org.apache.ibatis.session.SqlSession;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Repository;

@Repository
public class TradeDAO {
	@Autowired
	private SqlSession sqlSession;

	private final String NAMESPACE = "com.sun.app.trades.TradeDAO.";

	public int add(TradeDTO tradeDTO) throws Exception {
		return sqlSession.insert(NAMESPACE + "add", tradeDTO);
	}

	public int update(TradeDTO tradeDTO) throws Exception {
		return sqlSession.update(NAMESPACE + "update", tradeDTO);
	}
}
